+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [4.0.1] 2026-04-03
+
+#### Fixed
+- Sidebar optimization: steps that will be auto-skipped (via config/flags/env) are hidden
+  from the sidebar immediately after account selection, instead of showing as pending dots until the
+  wizard sequentially reaches each one.
+
 ## [4.0.0] 2026-04-02
 
 #### Added
@@ -138,7 +145,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Initial release.
 
-[Unreleased]: https://github.com/YouSysAdmin/jc2aws/compare/v4.0.0...HEAD
+[Unreleased]: https://github.com/YouSysAdmin/jc2aws/compare/v4.0.1...HEAD
+[4.0.1]: https://github.com/YouSysAdmin/jc2aws/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/YouSysAdmin/jc2aws/compare/v3.0.1...v4.0.0
 [3.0.1]: https://github.com/YouSysAdmin/jc2aws/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/YouSysAdmin/jc2aws/compare/v2.2.0...v3.0.0